30.3.Styling Emails with Encore & Sass Part 1{Creating the Email Entry}
Creating the Email Entry
Now open up the email layout file: templates/email/emailBase.html.twig. When we used inline_css(), we pointed it at the foundation-emails.css file and the email.css file. But now... we only really need to point it at email.scss... because, in theory, that will include the styles from both files.
The problem is that this is now a Sass file... and inline_css only works with CSS files: we can't point it at a Sass file and expect it transform the Sass into CSS. And even if it were a CSS file, the @import won't work unless we process this through Encore.
So here's the plan: we're going to pretend that email.scss is just an ordinary CSS file that we want to include on some page on our site. Open up webpack.config.js. Whenever we have some page-specific CSS or JS, we add a new entry for it. In this case, because we don't need any JavaScript, we can add a "style" entry. Say .addStyleEntry() - call the entry, how about, email, and point it at the file: ./assets/css/email.scss.
To get Webpack to see the updated config, in the terminal, press Ctrl+C to stop Encore and restart it:

yarn dev --watch
And... it builds! Interesting: the email entrypoint dumped two CSS files. Let's look at the public/build directory. Yep: email.css and also this vendors~email.css.
This is thanks to an optimization that Wepback Encore makes when you use splitEntryChunks()... which you can learn all about in our Encore tutorial. But the basic point is that if we want all of the CSS from the built email.scss file, we need to include both email.css and vendor~email.css.
Ok, easy, right? In the template, we could load the source of vendor~email.css and email.css. The problem is that Webpack splits the files in a very dynamic fashion: if it finds a more efficient way to split the files tomorrow - maybe into three files - it will! Plus, when we do our production build, the files will include a dynamic hash in their filename - like email.123abc.css.
So... we need to do a bit more work to reliably load this stuff through inline_css(). Let's do that next with a custom Twig function.