31.1.Processing Encore Files through inline_css()
We just used Encore to build an email.scss file that we want to process through inline_css() to style our emails. The problem is that, instead of building just one email.css file in public/build, it split it into two for performance reasons. That wouldn't be a problem, except that the way Webpack splits the files might change over time - we can't guarantee that it will always be these two files. To make matters worse, an Encore production build will add a dynamic "hash" to every file - like email.123abc.css.
Basically... pointing inline_css() directly at these two files... isn't going to work.

How Dynamic Files are Normally Rendered
This is why, in base.html.twig we simply use encore_entry_link_tags() and it takes care of everything. How? Behind the scenes, it looks in the public/build/ directory for an entrypoints.json file that Encore builds. This is the key: it tells us exactly which CSS and JS files are needed for each entrypoint - like app. Or, for email, yep! It contains the two CSS files.
The problem is that we don't want to just output link tags. We actually need to read the source of those files and pass that to inline_css().