31.3.Processing Encore Files through inline_css(){Adding a publicDir Binding}
Adding a publicDir Binding
We could hardcode the path to the public/ directory right here. But instead, let's set up a new "binding" that we can pass through the constructor. Open up config/services.yaml. In our Symfony Fundamentals Course, we talk about how the global bind below _defaults can be used to allow scalar arguments to be autowired into our services. Add a new one: string $publicDir set to %kernel.project_dir% - that's a built-in parameter - /public.
This string part before $publicDir is optional. But by adding it, we're literally saying that this value should be passed if an argument is exactly string $publicDir. Being able to add the type-hint to a bind is a new feature in Symfony 4.2. We didn't use it on the earlier binds... but we could have.
Back in AppExtension, add the string $publicDir argument. I'll hit "Alt + Enter" and go to "Initialize fields" to create that property and set it.
Down in the method, we can say $source .= file_get_contents($this->publicDir.$file) - each $file path should already have a / at the beginning. Finish the method with return $source.
Whew! Let's try this! We're already running Encore... so it already dumped the email.css and vendors~email.css files. Ok, let's go send an email. I'll hit back to get to the registration page, bump the email, type any password, hit register and... wow! No errors! Over in Mailtrap... nothing here... Of course! We refactored to use Messenger... so emails are not sent immediately!
By the way, if that annoys you in development, there is a way to handle async messages immediately while coding. Check out the Messenger tutorial.
Let's start the worker and send the email. I'll open another tab in my terminal and run:

php bin/console messenger:consume -vv
Message received... and... message handled. Go check it out! The styling look great: they're inlined and coming from a proper Sass file.
And... we've made it to the end! You are now an email expert... I mean, not just a Mailer expert... we really dove deep. Congrats!
Go forth and use your great power responsibly. Let us know what cool emails you're sending... heck... you could even send them to us... and, as always, we're here to help down in the comments section.

Alright friends, seeya next time!